use f32 as standard, allow switching to f64

+++ examples/noisy-poly.rs
@@ -5,7 +5,7 @@ use esopt::*;
 use rand::prelude::*;
 
 const DEGREE:usize = 3;
-const REGULARIZE:f64 = 0.025; //own L1 reg. factor
+const REGULARIZE:Float = 0.025; //own L1 reg. factor
 
 
 fn main()
@@ -51,7 +51,7 @@ fn main()
 }
 
 
-fn calc_f(factors:&[f64], x:f64) -> f64
+fn calc_f(factors:&[Float], x:Float) -> Float
 {
     let mut result = factors[0];
     let mut current_pow = 1.0;
@@ -63,7 +63,7 @@ fn calc_f(factors:&[f64], x:f64) -> f64
     result
 }
 
-fn print_f(fct:&[f64])
+fn print_f(fct:&[Float])
 {
     let mut str = String::new();
     for i in 0..fct.len()
@@ -79,12 +79,12 @@ fn print_f(fct:&[f64])
 #[derive(Clone)]
 struct PolynomeEval
 {
-    target:Vec<(f64,f64)>, //points that should be matched by the polynomial function
+    target:Vec<(Float,Float)>, //points that should be matched by the polynomial function
 }
 
 impl PolynomeEval
 {
-    pub fn new(points:Vec<(f64,f64)>) -> PolynomeEval
+    pub fn new(points:Vec<(Float,Float)>) -> PolynomeEval
     {
         PolynomeEval { target: points }
     }
@@ -93,7 +93,7 @@ impl PolynomeEval
 impl Evaluator for PolynomeEval
 {
     //evaluate as inverted mean absolute error to target (we want to minimize instead of maximize)
-    fn eval_train(&self, params:&[f64], _:usize) -> f64
+    fn eval_train(&self, params:&[Float], _:usize) -> Float
     {
         let mut rng = thread_rng();
         //calculate mean absolute error of random batch of points -> seems to accept outliers
@@ -106,14 +106,14 @@ impl Evaluator for PolynomeEval
             let error = point.1 - calc_f(params, point.0);
             mae += error.abs();
         }
-        mae /= n as f64;
+        mae /= n as Float;
         //regularize using L1
         let mut l1 = 0.0;
         for val in params.iter()
         {
             l1 += val.abs();
         }
-        l1 /= params.len() as f64;
+        l1 /= params.len() as Float;
         l1 *= REGULARIZE;
         //return
         -(mae + l1)
@@ -121,7 +121,7 @@ impl Evaluator for PolynomeEval
     
     //evaluate as inverted mean absolute error to target (we want to minimize instead of maximize)
     //here all data is used for deterministic progess monitoring
-    fn eval_test(&self, params:&[f64]) -> f64
+    fn eval_test(&self, params:&[Float]) -> Float
     {
         //calculate mean absolute error
         let mut mae = 0.0;
@@ -131,14 +131,14 @@ impl Evaluator for PolynomeEval
             let error = point.1 - calc_f(params, point.0);
             mae += error.abs();
         }
-        mae /= self.target.len() as f64;
+        mae /= self.target.len() as Float;
         //regularize using L1
         let mut l1 = 0.0;
         for val in params.iter()
         {
             l1 += val.abs();
         }
-        l1 /= params.len() as f64;
+        l1 /= params.len() as Float;
         l1 *= REGULARIZE;
         //return
         -(mae + l1)
